requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
Jinja2>=3.1.2
PyYAML>=6.0.1
python-json-logger>=2.0.7
beautifulsoup4>=4.12.3
lxml>=5.2.2
cssutils>=2.9.0
Pillow>=10.3.0
brotli>=1.1.0
# optional for runtime metrics:
playwright>=1.47.0

